+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - Normalized error handling for plugin.Open and Plugin.Lookup
+- Support for symbol defaults and searching over multiple plugins
 
 ## [v0.0.1]
 - Initial creation